lib/mito/utilities: move utilities to a separate header
