+++ lib/mito/utilities/forward.h
@@ -31,31 +31,6 @@ namespace mito::utilities {
     // requires ReferenceCountedObject<typename sharedResourceT::resource_type>
     class Repository;
 
-    // TOFIX: move these guys to a {utilities.h} header?
-    // segmented container iterator
-    // equality
-    template <class SegmentedContainerT>
-    constexpr auto operator==(
-        const SegmentedContainerIterator<SegmentedContainerT> & it1,
-        const SegmentedContainerIterator<SegmentedContainerT> & it2) noexcept -> bool;
-
-
-    // and not
-    template <class SegmentedContainerT>
-    constexpr auto operator!=(
-        const SegmentedContainerIterator<SegmentedContainerT> & it1,
-        const SegmentedContainerIterator<SegmentedContainerT> & it2) noexcept -> bool;
-
-    // hash function for shared pointers
-    // Note that two pointers pointing to the same cell collapse on the same hashed value
-    template <class sharedPointerT>
-    struct hash_function {
-        size_t operator()(const sharedPointerT & item) const
-        {
-            // reinterpret the address of the pointed handle as a {size_t} and return it
-            return reinterpret_cast<size_t>(item.id());
-        }
-    };
 }
